refactor: 수평/수직방향 MainScreen을 별도의 파일로 분리
